+++ app/src/main/java/io/github/dansager/travelplanner/data_structures/Trip.java
@@ -12,7 +12,7 @@ public class Trip {
     private String name = "";
     private DateTime startDate;
     private DateTime endDate;
-    private int MoneySpent = 0;
+    private double MoneySpent = 0;
     private List<Expense> list = new ArrayList<Expense>();
 
     public Trip(String name,DateTime startDate,DateTime endDate) {
@@ -25,7 +25,7 @@ public class Trip {
         return name = s;
     }
 
-    public int setMoneySpent (int i) {
+    public double setMoneySpent (double i) {
         return MoneySpent = i;
     }
 
@@ -36,7 +36,7 @@ public class Trip {
         return endDate = d;
     }
 
-    public int getMoneySpent() {
+    public double getMoneySpent() {
         return MoneySpent;
     }
 
@@ -57,14 +57,5 @@ public class Trip {
 
     public void addToList (Expense e) {
         list.add(e);
-//        Collections.sort(list, new Comparator<Expense>() {
-//            @Override
-//            public int compare(Expense expense1, Expense expense2) {
-//                DateTime d1 = new DateTime(expense1.getStartDate());
-//                DateTime d2 = new DateTime(expense2.getStartDate());
-//                return d1.compareTo(d2);
-//
-//            }
-//        });
     }
 }
